Fix: Sorting
By using s-join instead of org-agenda-finalize-entries, which does its
own sorting.  We don't need anything it does, at least not now, and if
we ever do, we'll probably want to reimplement it, anyway.

HOWEVER, this breaks org-super-agenda, which uses advice to that
function to group items.  So we need to either figure out another
approach (which I have been unable to do so far, trying cl-letf,
cl-flet, and advice-add with unwind-protect), or make this depend on
org-super-agenda.