Add: (org-ql-view--format-element) Set org-agenda-type to search
Some org-agenda commands check the org-agenda-type text property to
ensure that the command is valid to run in that type of agenda.
Notably this fixes running commands org-agenda-schedule and
org-agenda-deadline.

Setting the type to "search" prohibits some of the functions that are
still not working in these buffers, like org-agenda-do-date-earlier.

Fixes #35.