Tidy: Use alternatives to map-put
Emacs 27 marks map-put as obsolete and renames it to map-put!, which
causes byte-compilation warnings (sigh...isn't that the definition of
useless churn?).  Of course, map-put! doesn't exist in Emacs 26, so we
can't just use that.  I'm beginning to think that it will never be
possible to make the linting clean.