SIGNIFICANT speedup of org-ql-select (requires customization!)
With this change, and customizing org-ql-open-buffers to it's
non-default value will give HUGE performance improvements to org-ql
queries when files are queried that aren't already loaded into
buffers.

This change does not modify any defaults and org-ql will work just as
before (and as slow) until org-ql-open-buffers is modified by the
user.  This is due to backwards compatability.

* org-ql.el (org-ql-select): Don't always read files into buffers,
  depending on user customization.
  (org-ql--prepair-work-buffer, org-ql--work-buffer)
  (org-ql--init-work-buffer): Helper functions to work with new
  hidden work-buffer.
  (org-ql--work-buffer-name): New variable for name of work-buffer.
  (org-ql-open-buffers): User customization to choose whether org-ql
  should open all files in separate buffers or not.