Handle the case when org-use-property-inheritance is a list of strings
When org-use-property-inheritance is a list, then the query

(property "FOO" "BAR")

would bail out with:

org-ql--byte-compile-warning: Invalid Org QL query: "Invalid Org QL
query: \"‘\\\"BLAH\\\"’ is a malformed function\", :warning", :error

(property "FOO") would still work though.

So, if org-use-property-inheritance is a list, cast it to
the (default) boolean value for this variable: nil.

Use 'selective' inheritance when org-use-property-inheritance is a list

This symbol is mentioned in the org-entry-get documentation.