Remove mentions of removed functions from v0.7 subheading in Changelog.
These commands were removed in 5768c685adc7c6e85c17cfec358aa8a0e368310b,
but they were still mentioned in the release that was going to
incorporate those two functions, v0.7.

This commit removes the mention of those two functions from the v0.7
subheading in the changelog. Note that `org-ql-find-path` was
re-introduced in v0.8, so it remains under that subheading as a separate
addition.